+++ packages/reactivity-core/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @conterra/reactivity-core
 
+## v0.3.0 (unreleased)
+
+-   Rename `isReactive()` to `isReadonlyReactive()`
+-   Rename `isWritableReactive()` to `isReactive()`
+
 ## v0.2.0
 
 -   New feature: `reactiveStruct`, a utility to create simple reactive classes.